% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/because_loo.R
\name{because_loo_compare}
\alias{because_loo_compare}
\title{Compare Models Using LOO-CV}
\usage{
because_loo_compare(...)
}
\arguments{
\item{...}{Two or more \code{loo} objects from \code{because_loo()}.}
}
\value{
A comparison table ranking models by expected out-of-sample predictive accuracy.
}
\description{
Wrapper for \code{loo::loo_compare()} to compare multiple Because models.
}
\examples{
\dontrun{
  loo1 <- because_loo(fit1)
  loo2 <- because_loo(fit2)
  because_loo_compare(loo1, loo2)
}

}
