% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/because_random.R
\name{expand_nesting_syntax}
\alias{expand_nesting_syntax}
\title{Expand Nesting Syntax in Random Effects}
\usage{
expand_nesting_syntax(rhs)
}
\arguments{
\item{rhs}{Character string, RHS of random effects formula}
}
\value{
Expanded character string
}
\description{
Converts lme4-style nesting (1|A/B) to (1|A) + (1|A:B)
}
\keyword{internal}
