% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_helpers.R
\name{expand_polynomial_formula}
\alias{expand_polynomial_formula}
\title{Expand Polynomial Terms in Formula}
\usage{
expand_polynomial_formula(formula, poly_terms)
}
\arguments{
\item{formula}{Formula object}

\item{poly_terms}{List of polynomial term info from extract_polynomial_terms}
}
\value{
Modified formula with internal names
}
\description{
Replaces I(var^power) with internal variable names
}
\keyword{internal}
