% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mag_helpers.R
\name{extract_bidirected_edges}
\alias{extract_bidirected_edges}
\title{Extract bidirected edges (induced correlations) from MAG}
\usage{
extract_bidirected_edges(mag)
}
\arguments{
\item{mag}{MAG adjacency matrix from DAG.to.MAG}
}
\value{
List of variable pairs with induced correlations
}
\description{
Extract bidirected edges (induced correlations) from MAG
}
\keyword{internal}
