% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_definitions.R
\name{family_definitions}
\alias{family_definitions}
\title{Family Definition Generics and Default Methods}
\description{
S3 generics and methods for distribution families in because.
This enables custom distributions to be added by defining S3 methods.
}
