% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_helpers.R
\name{generate_polynomial_jags}
\alias{generate_polynomial_jags}
\title{Generate JAGS Code for Polynomial Transformations}
\usage{
generate_polynomial_jags(poly_terms)
}
\arguments{
\item{poly_terms}{List of polynomial term info}
}
\value{
Character vector of JAGS code lines
}
\description{
Creates deterministic nodes for polynomial terms
}
\keyword{internal}
