% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_data.R
\name{infer_variable_level}
\alias{infer_variable_level}
\title{Infer Variable Level}
\usage{
infer_variable_level(var, levels)
}
\arguments{
\item{var}{Character, variable name}

\item{levels}{List mapping variable names to level names}
}
\value{
Character, level name
}
\description{
Determine which hierarchical level a variable belongs to
}
\keyword{internal}
