% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phybase_compare.R
\name{phybase_compare}
\alias{phybase_compare}
\title{Compare PhyBaSE Models using WAIC}
\usage{
phybase_compare(..., models = NULL, sort = TRUE)
}
\arguments{
\item{...}{Fitted model objects of class \code{"phybase"}.}

\item{models}{Optional list of models. If provided, \code{...} is ignored.}

\item{sort}{Logical. If \code{TRUE} (default), sort models by WAIC (lowest to highest).}
}
\value{
A class \code{"phybase_comparison"} object containing a data frame with:
  \item{WAIC}{The WAIC value (lower is better).}
  \item{SE}{Standard error of WAIC.}
  \item{p_waic}{Effective number of parameters.}
  \item{dWAIC}{Difference in WAIC relative to the best model.}
  \item{dSE}{Standard error of the difference relative to the best model.}
  \item{weight}{Akaike-like weight (probability that this model is the best among the set).}
}
\description{
Compares multiple fitted PhyBaSE models using the Widely Applicable Information Criterion (WAIC).
Models are ranked from best (lowest WAIC) to worst.
}
\details{
The function extracts WAIC results from each model. If a model does not have
WAIC calculated, it gives a warning and excludes it or tries to calculate it
(currently assumes WAIC is present).

The standard error of the difference (\code{dSE}) is calculated as:
\code{sqrt(n * var(pointwise_waic_diff))}.

Weights are calculated as:
\code{exp(-0.5 * dWAIC) / sum(exp(-0.5 * dWAIC))}.
}
\examples{
\dontrun{
  fit1 <- phybase_run(..., WAIC = TRUE)
  fit2 <- phybase_run(..., WAIC = TRUE)

  comp <- phybase_compare(fit1, fit2)
  print(comp)
}
}
\references{
Vehtari, A., Gelman, A. & Gabry, J. (2017). Practical Bayesian model evaluation
using leave-one-out cross-validation and WAIC. Statistics and Computing, 27(5), 1413-1432.
}
