% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phybase_run.R
\name{phybase_run}
\alias{phybase_run}
\title{Run a Phylogenetic Bayesian Structural Equation model (PhyBaSE)}
\usage{
phybase_run(
  data,
  tree,
  equations,
  monitor = NULL,
  n.chains = 3,
  n.iter = 2000,
  n.burnin = floor(n.iter/2),
  n.thin = max(1, floor((n.iter - n.burnin)/1000)),
  DIC = TRUE,
  WAIC = FALSE,
  n.adapt = 100,
  quiet = FALSE,
  dsep = FALSE
)
}
\arguments{
\item{data}{A named list of data for JAGS.}

\item{tree}{A single phylogenetic tree of class \code{"phylo"} or a list of trees of class \code{"multiPhylo"} (i.e., a list of \code{"phylo"} objects). If multiple trees are provided, phylogenetic uncertainty is incorporated by sampling among them during model fitting.}

\item{equations}{A list of model formulas describing the structural equation model.}

\item{monitor}{Optional character vector of parameter names to monitor. If \code{NULL}, parameters will be selected automatically based on model structure.}

\item{n.chains}{Number of MCMC chains (default = 3).}

\item{n.iter}{Total number of MCMC iterations (default = 2000).}

\item{n.burnin}{Number of burn-in iterations (default = n.iter / 2).}

\item{n.thin}{Thinning rate (default = max(1, floor((n.iter - n.burnin) / 1000))).}

\item{DIC}{Logical; whether to compute DIC using \code{dic.samples()} (default = TRUE).}

\item{WAIC}{Logical; whether to sample values for WAIC and deviance (default = FALSE).}

\item{n.adapt}{Number of adaptation iterations (default = 100).}

\item{quiet}{Logical; suppress JAGS output (default = FALSE).}

\item{dsep}{Logical; if \code{TRUE}, monitor only the first beta in each structural equation (used for d-separation testing).}
}
\value{
A list of class \code{"phybase"} with model output and diagnostics.
}
\description{
Run a Phylogenetic Bayesian Structural Equation model (PhyBaSE)
}
