% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phybase_waic.R
\name{phybase_waic}
\alias{phybase_waic}
\title{Calculate WAIC for a PhyBaSE Model}
\usage{
phybase_waic(model, n.iter = 2000, n.burnin = 500, n.thin = 10)
}
\arguments{
\item{model}{A fitted model object of class \code{"phybase"} returned by \code{\link{phybase_run}}.}

\item{n.iter}{Number of iterations for WAIC calculation. Default is 2000.}

\item{n.burnin}{Number of burn-in iterations. Default is 500.}

\item{n.thin}{Thinning interval. Default is 10.}
}
\value{
A named vector containing:
\item{waic}{The calculated WAIC value.}
\item{p_waic}{The effective number of parameters (penalty term).}
}
\description{
Calculates the Widely Applicable Information Criterion (WAIC) for a fitted PhyBaSE model.
}
\details{
This function uses the \code{dic} module in JAGS to monitor \code{WAIC} and \code{deviance}.
The WAIC is calculated as: \code{WAIC = mean(deviance) + p_waic}.
}
\examples{
\dontrun{
  fit <- phybase_run(data, tree, equations)
  waic_res <- phybase_waic(fit, n.iter = 1000)
  print(waic_res)
}

}
