% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.because.R
\name{summary.because}
\alias{summary.because}
\title{Summary for Because Model}
\usage{
\method{summary}{because}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class \code{"because"}.}

\item{...}{Additional arguments passed to \code{\link[coda]{summary.mcmc}}.}
}
\value{
A summary object containing statistics for the monitored parameters.
  If \code{dsep = TRUE} was used in \code{because}, the summary focuses on
  the conditional independence tests.
}
\description{
Summarizes the output of a Because model run.
}
