% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.phybase.R
\name{summary.phybase}
\alias{summary.phybase}
\title{Summary for PhyBaSE Model}
\usage{
\method{summary}{phybase}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class \code{"phybase"}.}

\item{...}{Additional arguments passed to \code{\link[coda]{summary.mcmc}}.}
}
\value{
A summary object containing statistics for the monitored parameters.
  If \code{dsep = TRUE} was used in \code{phybase_run}, the summary focuses on
  the conditional independence tests.
}
\description{
Summarizes the output of a PhyBaSE model run.
}
